% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.aging.R
\name{aging_msm}
\alias{aging_msm}
\title{Aging Module}
\usage{
aging_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns \code{dat} after updating the nodal attribute
\code{age}.
}
\description{
Module for aging over time for active nodes in the population.
}
