% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_core_attr_mig}
\alias{append_core_attr_mig}
\title{Function to append core attributes}
\usage{
append_core_attr_mig(dat, at, n.new)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{at}{current time step}

\item{n.new}{the number of new nodes to initiate with core attributes}
}
\value{

}
\description{
Function to append core attributes
}
\details{
This function is based on \link[EpiModel]{append_core_attr}. I added migrationTime
as another core attribute.
}
