% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{convert_branches}
\alias{convert_branches}
\title{Title Scale branch length from a phylogeentic tree}
\usage{
convert_branches(tree, scale = 1/365)
}
\arguments{
\item{tree}{a phylo object}

\item{scale}{numeric value to convert branch length}
}
\value{

}
\description{
Title Scale branch length from a phylogeentic tree
}
\examples{
# TO DO!
}
