% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_inf_csv}
\alias{create_inf_csv}
\title{Create transmission matrix csv}
\usage{
create_inf_csv(tm, time_tr, prefix = NULL)
}
\arguments{
\item{tm}{Transmission matrix as returned using the function \link[EpiModel]{get_transmat}}

\item{time_tr}{Time of transmission for seeds. The other time of transmission
should be provided in the tm dataframe.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{

}
\description{
Create a transmission matrix file to be used with the
   VirusTreeSimulator. VirusTreeSimulator requires that seeds are included
   in the file.
}
\details{
If a prefix is not provided, csv file will be saved as inf.csv
}
\examples{
TO DO
}
