% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_sample_csv}
\alias{create_sample_csv}
\title{Get sample csv file}
\usage{
create_sample_csv(tm, time_seq, seq_count, prefix = NULL)
}
\arguments{
\item{tm}{Transmission matrix as returned using the function \link[EpiModel]{get_transmat}}

\item{time_seq}{Vector for sample time. If one value is provided, it will be replicated
to the size of samples in the transmission matrix (tm).}

\item{seq_count}{The number of sequences per sample in the transmission matrix.
If one value is provided, it will be replicated to the size of samples in
the transmission matrix (tm).}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{

}
\description{
Get sample csv file
}
\details{
If a prefix is not provided, csv file will be saved as sample.csv
}
\examples{
To Do
}
