% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.departure.R
\name{departure_mig}
\alias{departure_mig}
\title{Depature Module}
\usage{
departure_mig(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the updated \code{dat} object accounting for deaths.
}
\description{
Module function for simulating both general and disease-related
             departures, including deaths, among population members.
}
\details{
Deaths are divided into two categories: general deaths, for which demographic
data on age-specific mortality rates applies; and AIDS-mortality rate.
}
\keyword{module}
\keyword{msm}
