% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.departure.R
\name{departure_mig3}
\alias{departure_mig3}
\title{Depature Module}
\usage{
departure_mig3(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}
}
\value{
This function returns the updated \code{dat} object accounting for deaths.
}
\description{
Module function for simulating both general and disease-related
             departures, including deaths, among population members.
             In this function we use two departure rates: one for population
             1 and another for population 2.
}
\details{
Deaths are divided into two categories: general deaths, for which demographic
data on age-specific mortality rates applies; and AIDS-mortality rate.
}
\keyword{module}
\keyword{msm}
