% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model0_net.R
\name{generate_model0_net}
\alias{generate_model0_net}
\title{Generate a simple HIV epidemic model for use by phydynR}
\usage{
generate_model0_net(art_start = 2005)
}
\value{
A list with elements
* dm : a demographic process model
* parms: a default list of parameters
* x0 : a default vector of start conditions
}
\description{
* src : global reservoir
* i : infected and infectious
* z : infected and on ART
Transmission occurs at a rate beta(t) i(t) where beta is a linear
interpolating spline. The treatment rate i->z is initially zero but
increases linearly to 2015
}
