% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.transmat.phylo_mod}
\alias{get.transmat.phylo_mod}
\title{Get transmat infection tree for each seed}
\usage{
get.transmat.phylo_mod(x, vertex.exit.times)
}
\arguments{
\item{x}{An object of class \code{"transmat"}, the output from
\code{\link{get_transmat}}.}

\item{vertex.exit.times}{optional numeric vector providing the time of
departure of vertices, to be used to scale the lengths of branches
reaching to the tips. Index position on vector corresponds to network
id. NA indicates no departure, so branch will extend to the end of the
tree.}
}
\description{
Separate a \code{transmat} object into different seedling
      event.
}
\details{
Converts a \code{\link{transmat}} object containing information about the
history of a simulated infection into a separated transmission matrix by
seedling event.

This function was created using as base the function
\code{\link{as.phylo.transmat}}
}
