% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cd4s}
\alias{get_cd4s}
\title{Get CD4s by node IDs}
\usage{
get_cd4s(IDPOP, df_actives, df_departures)
}
\arguments{
\item{df_actives}{dataframe of IDs that were active at the end of a simulation
and stages of HIV infection.}

\item{df_departures}{dataframe of IDs that departed the network before end of
simulation and stages of HIV infection}
}
\value{

}
\description{
Get CD4s by node IDs
}
