% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{get_epi_data}
\alias{get_epi_data}
\title{Get epi data for simulation runs}
\usage{
get_epi_data(list_dirs)
}
\arguments{
\item{list_dirs}{list of directories containing data}
}
\value{
A list of dataframes for each type of simulations.
}
\description{
Get epi data for simulation runs
}
\details{
This function will go over each simulation run saved as RDS format,
  convert it to a dataframe and assign as simulation number the number of the
  respective run. This function was designed when running more than 1
  simulation independently in the cluster, to merge all the runs together for
  plotting purposes.
}
