% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{get_newTip_names}
\alias{get_newTip_names}
\title{Get tip names based on duplicated rows.}
\usage{
get_newTip_names(dup_tip, el_ori)
}
\arguments{
\item{dup_tip}{Vector of duplicated tips}

\item{el_ori}{Matrix of edge list and information on origin or migration}
}
\value{
A vector of the duplicated vertex in the form of ID_origin or ID_migrant.
   Based on the oldest time of duplicated rows (see \code{\link{oldest_time}}),
   it will return the tip name that should be in the phylogenetic tree.
   I used the origin or migrant information observed at the time of
   transmission.
}
\description{
This is a support function for the main function
   \code{\link{toPhylo_transmatOrigin}}.
}
