% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{get_rate}
\alias{get_rate}
\title{Get rate based on linear interpolation}
\usage{
get_rate(init_date, times, rates, at)
}
\arguments{
\item{init_date}{Object of class Date. Initial date for the start of simulations
in the form of "YEAR-MONTH-DAY".}

\item{times}{Vector of times for interpolation}

\item{rates}{Vector os respective rates (related to times)}

\item{at}{time step in days}
}
\value{
rate based on linear interpolation. See \code{\link[stats]{approx}}.
}
\description{
Get rate based on linear interpolation
}
\examples{
# times <- c(2000:2019)
# rates <- c(1.445432e-06, 1.455312e-06, 1.414593e-06, 1.346566e-06,
#            1.329516e-06, 1.281034e-06, 1.445432e-06, 1.455312e-06,
#             1.414593e-06, 1.346566e-06, 1.329516e-06, 1.281034e-06,
#             1.288603e-06, 1.250734e-06, 1.265186e-06, 1.234629e-06,
#              1.193328e-06, 1.133912e-06, 1.099263e-06, 1.052727e-06)
# init_date <- as.Date("1980-01-01")
# new_rate <- get_rate(init_date, times, rates, at = 10000)
}
