% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{get_tip_cherry}
\alias{get_tip_cherry}
\title{Get tip names for cherries in a phylogenetic tree}
\usage{
get_tip_cherry(phy)
}
\arguments{
\item{phy}{object of class phylo}
}
\value{
list of pairs of tip_names that are cherries
}
\description{
Get tip names for cherries in a phylogenetic tree
}
