% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{get_tip_names}
\alias{get_tip_names}
\title{Convert transmat infection tree into a phylogenetic tree}
\usage{
get_tip_names(
  x,
  format = "migrant",
  by_areas = "region",
  max_value = NULL,
  tips_only = TRUE
)
}
\arguments{
\item{x}{An object of class \code{"transmat"}, the output from
\code{\link{get_transmat}}.}

\item{format}{If format = "origin" return tip in the form of ID_global or ID_region.
If format = "migrant" return tip in the form of ID_1, ID_2, ID_21, ID_12}

\item{by_areas}{by_areas = "all" if interested in all seeds, or by_areas = "region"
if interested in seeds for region only.}

\item{max_value}{max value for comparison to select the seed values. It will
select only seed values less than max_value for seeds from region only.}

\item{tips_only}{tips_only = TRUE returns only the tip names of the phylogenetic tree,
tips_only = FALSE returns the phylogenetic tree.}
}
\description{
Converts the edgelist matrix in the \code{transmat} object into
             a \code{phylo} object by doing the required reordering and
             labeling. This is a specific function that requires simulations
             using EpiModel using the package HIVepisim (this package)
}
\details{
This code was build using the algorithm as in \code{\link{as.phylo.transmat}}.
The difference is that if using format = "origin", tip names of the phylogenetic
tree will have appended to their names their origin at the moment infection
happened (region or global). If using format = "migrant", tip names of the
phylogenetic tree will have appended to their names if they are a migrant
or not at the time of infection.
Migrant values can take the value of 1 if vertex is from region,
2 if vertex is from global,
21 if vertex is from global that migrated to region and
12 if vertex is from region that migrated to global.

Converts a \code{\link{transmat}} object containing information about the
history of a simulated infection into a \code{\link{phylo}} object
representation suitable for plotting as a tree with
\code{\link[ape]{plot.phylo}}. Each infection event becomes a 'node'
(horizontal branch) in the resulting phylo tree, and each network vertex
becomes a 'tip' of the tree. The infection events are labeled with the vertex
id of the infector to make it possible to trace the path of infection.

The infection timing information is included to position the phylo-nodes,
with the lines to the tips drawn to the max time value +1 (unless
\code{vertex.exit.times} are passed in it effectively assumes all vertices
are active/alive until the end of the simulation).

If the transmat contains multiple infection seeds (there are multiple trees
with seperate root nodes) it will return a list of class 'multiPhylo', each
element of which is a phylo object.  See \code{\link[ape]{read.tree}}.

Note that in EpiModel versions <= 1.2.4, the phylo tree was constructed
differently, translating network vertices to both phylo-nodes and tips and
requiring 'collapse.singles' to prune it to an appropriate branching
structure.

This script will specifically work with the package HIVepisim.
}
