% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.hivprogress.R
\name{hivprogress_msm}
\alias{hivprogress_msm}
\title{Disease Progression Module}
\usage{
hivprogress_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object after updating the disease stage
of infected individuals.
}
\description{
Module function for HIV disease progression
}
\details{
HIV disease is divided into four stages: stage 0 or acute and early HIV infection,
acute rising, and stage 1 to stage 4 (stage 4 is AIDS) as explained in
Cori et al. 2015.
}
\references{
Cori A, Pickles M, van Sighem A, et al. CD4+ cell dynamics in untreated
HIV-1 infection: overall rates, and effects of age, viral load, sex and
calendar time. AIDS. 2015;29(18):2435-2446.
doi:10.1097/QAD.0000000000000854
}
