% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.hivtest.R
\name{hivtest_msm}
\alias{hivtest_msm}
\title{HIV Testing Module}
\usage{
hivtest_msm(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{last.neg.test},
\code{diag.status} and \code{diag.time} attributes. Summary statistics for
number of new diagnoses, total number of tests, and total number of negative
tests are calculated and stored on \code{dat$epi}.
}
\description{
Module function for HIV diagnostic testing of infected persons.
}
\keyword{module}
\keyword{msm}
