% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.hivtrans.R
\name{hivtrans_mig}
\alias{hivtrans_mig}
\title{Transmission Module}
\usage{
hivtrans_mig(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}
}
\value{
For each new infection, the disease status, infection time, and related
HIV attributes are updated for the infected node. Summary statistics for
disease incidence overall are calculated and stored on \code{dat$epi}.
}
\description{
Stochastically simulates disease transmission given the current
             state of the discordand edgelist. This function should be used
             when there is migration.
}
\details{
This is the final substantive function that occurs within the time loop at
each time step. This function takes the discordant edgelist and calculates a
transmission probability for each row between dyads on the
network. After transmission events, individual-level attributes for the infected
persons are updated and summary statistics for incidence calculated.
}
