% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.hivtx.R
\name{hivtx_msm}
\alias{hivtx_msm}
\title{Treatment Module}
\usage{
hivtx_msm(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with updated \code{tx.status},
\code{tx.init.time}, \code{cuml.time.on.tx}, \code{cuml.time.off.tx} attributes.
Summary statistics for the average cumulative time on and off treatment
are calculated and stored on \code{dat$epi}.
}
\description{
Module function for anti-retroviral treatment (ART) initiation and
             adherence over time.
}
\details{
Persons enter into the simulation with one of three ART "patterns": never
tested, tested but never treated, and treated. This module initiates ART
for treatment naive persons in the latter type, and then cycles them on
and off treatment. ART initiation, non-adherence, and restarting are all
stochastically simulated based on binomial statistical models.
}
