% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_utils.R
\name{keep_row}
\alias{keep_row}
\title{Get index of rows to keep in trasmission matrix}
\usage{
keep_row(df, tree)
}
\arguments{
\item{tree}{object of class phylo}

\item{tm}{transmission matrix}
}
\value{

}
\description{
Get index of rows to keep in trasmission matrix
}
