% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{merge_trees}
\alias{merge_trees}
\title{Merge phylogenetic trees into a single tree}
\usage{
merge_trees(trees)
}
\arguments{
\item{trees}{list of phylogenetic trees}
}
\value{
merged trees as object of class phylo
}
\description{
Merge 2 or more phylogenetic trees into a single phylogenetic tree
}
