% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.migration.R
\name{migration2}
\alias{migration2}
\title{Migration Module}
\usage{
migration2(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}
}
\value{
This function updates the \code{attr} list with new attributes for each new
population member, and the \code{nw} objects with new vertices.
}
\description{
Module function for migrations between two populations
}
\details{
New migrations from both directions (from population 1 to 2; and
from population 2 to 1) will happen based on migration rates and population
size for each population stocastically determined with with draws from
Poisson distributions. For each new entry, a set of attributes is added for
that node, and the nodes are added onto the network objects.
Only attributes that are a part of the network model
formulae are updated as vertex attributes on the network objects.
}
\keyword{module}
\keyword{msm}
