% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{name_of_tips}
\alias{name_of_tips}
\title{Get tip names.}
\usage{
name_of_tips(dup_tip, row_int)
}
\arguments{
\item{dup_tip}{Vector of duplicated tips}

\item{row_int}{data.frame of the row of interest. The row in the transmission
matrix that have the oldest time step based on duplicated value.
See \code{\link{oldest_time}} and \code{\link{get_newTip_names}}.}
}
\value{
vector of ID_origin or ID_migrant.
}
\description{
This is a support function for the main function
   \code{\link{toPhylo_transmatOrigin}}.

Function that gest the name of the tip in the form of ID_origin or ID_migrant
because we can observe duplicates (when in the form ID_origin or ID_migrant),
we need to remove those from the phylogenetic tree.
}
