% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.nwupdate.R
\name{nwupdate_mig}
\alias{nwupdate_mig}
\title{Dynamic Network Updates}
\usage{
nwupdate_mig(dat, at)
}
\arguments{
\item{dat}{Master list object containing a full \code{networkDynamic} object
or networkLite edgelist (if using tergmLite), and other initialization
information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This function handles all calls to the network object contained
             on the master dat object handled in \code{netsim}..
}
