% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{oldest_time}
\alias{oldest_time}
\title{Get oldest times of duplicates.}
\usage{
oldest_time(dup_rows, get_last_time)
}
\arguments{
\item{dup_rows}{matrix of list of matrices of transmat
that have duplicated vertex IDs.}

\item{get_last_time}{vector of oldest time}
}
\description{
This is a support function for the main function
   \code{\link{toPhylo_transmatOrigin}}.
}
\details{
During the simulations, vertex can migrate from region to global
or global to region. When adding the information of migrant or origin to the
tip name, we will observe duplicated on vertex IDs because it can be in region
at one time point but later in global, for example.
This function will get the duplicated rows in the transmission matrix
in which a duplicated vertex is observed. It will then get the oldest observed
time for each duplicated vertex.
}
