% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_SDiego.R
\name{param_msm_sdiego}
\alias{param_msm_sdiego}
\title{Epidemic Model Parameters}
\usage{
param_msm_sdiego(
  netstats,
  time.unit = 7,
  stage_prog_rate1 = 1/((3.32 * 365)/7),
  stage_prog_rate2 = 1/((2.7 * 365)/7),
  stage_prog_rate3 = 1/((5.5 * 365)/7),
  f1 = 0.76,
  f2 = 0.19,
  f3 = 0.05,
  f4 = 0,
  hiv.test.rate = 0.01325,
  test.window.int = 21/7,
  tx.init.prob = 0.092,
  tx.halt.prob = 0.0102,
  tx.reinit.prob = 0.00066,
  trans.r = 1e-04,
  ws0 = 1,
  ws1 = 0.1,
  ws2 = 0.1,
  ws3 = 0.1,
  ws4 = 0.3,
  wc1 = 1,
  wc2 = 0.5,
  wc3 = 0.05,
  wr1 = 1,
  wr2 = 10,
  aids.mr = 1/((5.06 * 365)/7),
  a.rate = 0.00052,
  arrival.age = 18,
  ...
)
}
\arguments{
\item{netstats}{Target statistics and related network initialization data from
the standard ARTnet workflow.}

\item{time.unit}{Unit of time relative to one day.}

\item{stage_prog_rate1}{Rate per time unit spent in CD4+ compartment for stage 1 of HIV
infection (Cori et al. 2015).}

\item{stage_prog_rate2}{Rate per time unit spent in CD4+ compartment for stage 2 of HIV
infection (Cori et al. 2015).}

\item{stage_prog_rate3}{Rate per time unit spent in CD4+ compartment for stage 3 of HIV
infection (Cori et al. 2015).}

\item{f1}{Probability of an individual entering stage 1 of HIV infection after
initial HIV infection (acute and early HIV infection).}

\item{f2}{Probability of an individual entering stage 2 of HIV infection after
initial HIV infection (acute and early HIV infection).}

\item{f3}{Probability of an individual entering stage 3 of HIV infection after
initial HIV infection (acute and early HIV infection).}

\item{f4}{Probability of an individual entering stage 4 of HIV infection after
initial HIV infection (acute and early HIV infection).}

\item{hiv.test.rate}{Mean probability of HIV testing per week for
MSM.}

\item{test.window.int}{Length of the HIV test window period in weeks.}

\item{tx.init.prob}{Probability per time step that an MSM who has
tested positive will initiate treatment.}

\item{tx.halt.prob}{Probability per time step that an
MSM who have started treatment will stop treatment.}

\item{tx.reinit.prob}{Probability per time step that an
MSM who has stopped treatment will restart treatment.}

\item{trans.r}{Baseline for transmission rate per act.}

\item{ws0}{Risk ratio of an individual in acute and early HIV infection (stage 0)
will transmit to another individual.}

\item{ws1}{Risk ratio of an individual in stage 1 will transmit to an HIV negative
individual.}

\item{ws2}{Risk ratio of an individual in stage 2 will transmit to an HIV negative
individual.}

\item{ws3}{Risk ratio of an individual in stage 3 will transmit to an HIV negative
individual.}

\item{ws4}{Risk ratio of an individual in stage 4 will transmit to an HIV negative
individual.}

\item{wc1}{Risk ratio of an undiagnosed HIV + individual will transmit to an HIV
negative individual.}

\item{wc2}{Risk ratio of an diagnosed and untreated HIV + individual will
transmit to an HIV negative individual.}

\item{wc3}{Risk ratio of an diagnosed and treated HIV + individual will
transmit to an HIV negative individual.}

\item{wr1}{Risk ratio of individuals in risk group 1 will transmit to an HIV
negative individual.}

\item{wr2}{Risk ratio of individuals in risk group 2 will transmit to an HIV
negative individual.}

\item{aids.mr}{Mortality rate of persons in the AIDS stage who are currently
off ART.}

\item{a.rate}{Rate at which MSM enter the population.}

\item{arrival.age}{Age (in years) of new arrivals.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list object of class \code{param_msm}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\references{
Cori A, Pickles M, van Sighem A, et al.
CD4+ cell dynamics in untreated HIV-1 infection: overall rates, and effects
of age, viral load, sex and calendar time. AIDS. 2015;29(18):2435-2446.
doi:10.1097/QAD.0000000000000854
}
\keyword{msm}
