% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model0_net.R
\name{prepare_dated_trees}
\alias{prepare_dated_trees}
\title{Function that converts a dated phylogenetic tree into a
phydynR::DatedTree object}
\usage{
prepare_dated_trees(path_to_gene, path_to_sts)
}
\arguments{
\item{path_to_gene}{Path to where dated phylogenetic tree is saved}

\item{path_to_sts}{Path to where RData file containing sampling date}
}
\description{
Function that converts a dated phylogenetic tree into a
phydynR::DatedTree object
}
