% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.prevalence.R
\name{prevalence_mig}
\alias{prevalence_mig}
\title{Prevalence Calculations within Time Steps for 2 populations}
\usage{
prevalence_mig(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} object with an updated summary of current
attributes stored in \code{dat$epi}.
}
\description{
This module calculates demographic, transmission, and clinical
             statistics at each time step within the simulation.
}
\details{
This function establishes the summary statistic vectors for both
prevalence and incidence at time 1, and then calculates the prevalence
statistics for times 2 onward. Incidence statistics (e.g., number of new
infections or deaths) are calculated within the modules (arrival and departure)
as they depend on vectors that are not stored external to the module.
}
