% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{reorder_tip_names}
\alias{reorder_tip_names}
\title{Reorder tip names}
\usage{
reorder_tip_names(tip_names_migrant, tip_names_vts)
}
\arguments{
\item{tip_names_migrant}{Tip names in the form of ID_migrant}

\item{tip_names_vts}{Tip names as returned by using the VirusTreeSimulator.}
}
\value{

}
\description{
Function to reorder the tip names of a phylogenetic tree.
}
\details{
Vector of tip_names_migrant will be reordered according to
   vector tip_names_vts.
}
\examples{
#TO DO!
}
