% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{replace_dups}
\alias{replace_dups}
\title{Replace duplicates.}
\usage{
replace_dups(dup_tip, origNodes, tipNamesOri, tip_names_toInsert, tip_names)
}
\arguments{
\item{dup_tip}{Vector of duplicated tips}

\item{origNodes}{Vector of vertex IDs.}

\item{tipNamesOri}{Vector in the form of ID_origin or ID_migrant. This vector
contain the duplicated vertex when considering only the ID.}

\item{tip_names_toInsert}{Vector of the correct value of tip name to use in
the phylogenetic tree in the form of ID_origin or ID_migrant}

\item{tip_names}{Vector of tip names in the form of ID only. This vector will
contain the duplicated ID or IDs.}
}
\value{
Vector of tip names in the form of ID_origin or ID_migrant to use
   as tip.labels in the phylogenetic tree.
}
\description{
This is a support function for the main function
   \code{\link{toPhylo_transmatOrigin}}.

Replace duplicated in the form o ID_origin or ID_migrant by the correct tip
name.
}
