% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{sampler}
\alias{sampler}
\title{Sampler to sample parameters from a uniform distribution}
\usage{
sampler(n = 100, paramdf)
}
\arguments{
\item{n}{numeric for the number of samples. Default is set to 100.}

\item{paramdf}{<- dataframe for parameters names and values}
}
\value{
dataframe for sampled parameters.
}
\description{
Funtion that will sample parameter values from a uniform distribution based
on latin hypercube sampling.
}
\details{
This function will sample from a uniform distribution parameter
 values for initial population size for infected individuals (init.pop.param),
 act rate parameter (act.rate.param), and infection probability parameters
 (inf.prob.paramter)
 Boundaries for each parameter:
 init.pop.param: min = 100, max = 10000
 act.rate.param: min = , max =
 inf.prob.parameter: min = 0.01, max = 1
}
\examples{
#the example below will sample 100 parameter values for each b, initS, import,
and st from a uniform ditribution
parameters <- sampler()
}
