% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_utils.R
\name{sampler2}
\alias{sampler2}
\title{Sampler to sample parameters from a uniform distribution}
\usage{
sampler2(n = 100, paramdf)
}
\arguments{
\item{n}{numeric for the number of samples. Default is set to 100.}

\item{paramdf}{dataframe for parameters names and values}
}
\value{
dataframe for sampled parameters.
}
\description{
Funtion that will sample parameter values from a uniform distribution based
on latin hypercube sampling.
}
\details{
This function will sample from a uniform distribution 2 different
 parameter values of interest. Maximum and minimum values to sample from
 uniform distribution should be provided.
}
