% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{save_art}
\alias{save_art}
\title{Save time in which antiretroviral treatment (ART) started}
\usage{
save_art(dat, prefix = NULL)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{
it writes a csv file at a specific location at the computer
}
\description{
Whenever an individual initiates ART, this function will save
   the IDs of infected nodes and when ART started.
}
\details{
This function will save the time in which an individual initiated ART


If a prefix is not provided, csv file will be saved as ART_init.csv
}
