% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_departures}
\alias{save_departures}
\title{Save id and time of individuals that departure the network}
\usage{
save_departures(dat, departures, at, prefix = NULL)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{departures}{ids of departures}

\item{at}{Current time step.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{

}
\description{
It aims to save the ID and time of infected individual that
 departure the network via natural or HIV related cause.
}
\details{
If a prefix is not provided, csv file will be saved as departure_IDs.csv
}
\examples{
TO DO
}
