% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{save_diagnosis_time}
\alias{save_diagnosis_time}
\title{Save time in which an individual has been diagnosed as HIV positive}
\usage{
save_diagnosis_time(dat, prefix = NULL)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{

}
\description{
Whenever an individual has a positive HIV test, this
  function will save the ID of the individual and the time of diagnosis.
}
\details{
This function will save the time in which an individual has been diagnosed
as HIV positive.


If a prefix is not provided, csv file will be saved as diag_time.csv
}
