% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{save_nodes}
\alias{save_nodes}
\title{Save total number of nodes at each time step by attribute origin}
\usage{
save_nodes(dat, prefix = NULL)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{

}
\description{
This function will save the total number of nodes by attribute origin
}
\details{
If a prefix is not provided, csv file will be saved as total_nodes.csv
}
