% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_origin}
\alias{save_origin}
\title{Save origin of nodes in the network at final step}
\usage{
save_origin(dat, prefix = NULL)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{

}
\description{
At the final step of network simulation, it will save the IDs of
   infected nodes and their origin. Origin can take the value of region or
   global.
}
\details{
If a prefix is not provided, csv file will be saved as infected_origin.csv
}
\examples{
TO DO
}
