% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{save_r0}
\alias{save_r0}
\title{Save time and values for basic reproduction number}
\usage{
save_r0(dat, prefix = NULL)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{
it writes a csv file to the specified location in the computer
}
\description{
This function will save information for cumulative basic reproduction
   number.
}
\details{
If a prefix is not provided, csv file will be saved as r0.csv
}
