% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{save_track_origin}
\alias{save_track_origin}
\title{Save time and IDs of nodes and their location.}
\usage{
save_track_origin(dat, prefix = NULL)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{
it writes a csv file to the specified location in the computer
}
\description{
Whenever an individual change their origin because of migration,
  this function will save the information on track_origin.
}
\details{
This function will save information to get location of individual at certain
   time.


If a prefix is not provided, csv file will be saved as origin.csv
}
