% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{set_art_reinit}
\alias{set_art_reinit}
\title{Set time in which antiretroviral treatment (ART) was reinitiated}
\usage{
set_art_reinit(dat, at, IDs)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}

\item{IDs}{IDs of individuals who started ART.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{
the Master list object of network models (\code{dat})
}
\description{
Whenever an individual initiates ART, this function will save
   the IDs of infected nodes and when ART started.
}
\details{
This function will save the time in which an individual initiated ART
}
