% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{set_diagnosis_time}
\alias{set_diagnosis_time}
\title{Set time in which an individual has been diagnosed as HIV positive}
\usage{
set_diagnosis_time(dat, at, IDs)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}

\item{IDs}{IDs of individuals who have been diagonsed as HIV positive.}

\item{prefix}{Text for prefix to use when saving filename.}
}
\value{

}
\description{
Whenever an individual has a positive HIV test, this function
   will save the IDs of infected nodes and time of the positive HIV test.
}
\details{
This function will save the time in which an individual has diagnosed as
HIV positive


If a prefix is not provided, csv file will be saved as ART.csv
}
