% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMSM_population.R
\name{simulate_msm_pop}
\alias{simulate_msm_pop}
\title{Simulate population of MSM in San Diego}
\usage{
simulate_msm_pop(
  npop1,
  npop2,
  init_infect_pop2,
  init.pop1.param = 300.9307,
  act.rate.param = 0.22,
  tx.init.prob.param
)
}
\arguments{
\item{npop1}{Initial number of individuals in population 1 (susceptibles + infected)}

\item{npop2}{Initial number of individuals in population 2 (susceptibles + infected)}

\item{init_infect_pop2}{Initial number of infected individuals in population 2}

\item{init.pop1.param}{Initial number of infected individuals in population 1.
Fixed at 300.9307 (according the ECDC model)}

\item{act.rate.param}{Act rate parameter fixed at 0.22 per day (81 acts per year)}

\item{tx.init.prob.param}{Probability an MSM will initiated treatment per day
when diagnosed with HIV.}
}
\value{

}
\description{
Simulate population of MSM in San Diego
}
