% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMSM_population.R
\name{simulate_population}
\alias{simulate_population}
\title{Simulate population of MSM in San Diego}
\usage{
simulate_population(
  npop1,
  npop2,
  init_infect_pop2,
  init.pop1.param,
  act.rate.param,
  inf.prob.param
)
}
\arguments{
\item{npop1}{Initial number of individuals in population 1 (susceptibles + infected)}

\item{npop2}{Initial number of individuals in population 2 (susceptibles + infected)}

\item{init_infect_pop2}{Initial number of infected individuals in population 2}

\item{init.pop1.param}{Initial number of infected individuals in population 1}

\item{act.rate.param}{Act rate parameter}

\item{inf.prob.param}{Infection probability parameter}
}
\value{

}
\description{
Simulate population of MSM in San Diego
}
