% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_utils.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize data for statistics calculated for infector probability (W)}
\usage{
summarize_data(df)
}
\arguments{
\item{df}{dataframe containing data to summarize}
}
\value{
dataframe for median and quantiles for quantities of interest.
  see \code{link{validadeW}}
}
\description{
Summarize data for statistics calculated for infector probability (W)
}
