% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_utils.R
\name{summaryW2}
\alias{summaryW2}
\title{Title Validate infector probability calculation}
\usage{
summaryW2(sim, run, tm, W1, ID, MH, tree, prefix = NULL, labels = TRUE)
}
\arguments{
\item{sim}{Simulation number to know the parameter values that network was
simulated}

\item{run}{Run number}

\item{tm}{Transmission matrix}

\item{W1}{Infector probability calculated on true trees (all region tips)}

\item{ID}{ID name.}

\item{MH}{maximum height in which infector probabilities should be calculated}

\item{prefix}{prefix to save results. If prefix = NULL, results will be saved
on file W_on_true.csv.}

\item{W2}{Infector probability calculated on true trees (subset of region tips)}

\item{true1}{object of class phylo}

\item{true2}{object of class phylo}
}
\value{
a data frame of 1 row with the following values:
 1. Code = code indicating if W was calculated on true trees
 2. Maximum height = maximum height (MH) in which calculations were performed;
 3. Total tips = total number of tips in the phylogenetic tree;
 4. n_tips_region = total number of tips from region in the phylogenetic tree;
 5. n_tips_global = total number of tips from global in the phylogenetic tree;
 6. n_trans_W = total number of transmissions independent of infector probability
    values;
 7. n_trans_tm = total number of transmissions based on transmission matrix
    subset by MH;
 8. n_true_trans_W = total number of true transmission independent of infector
    probability (W) values. Here is the true transmission when compared to the
    transmission matrix;
 9. n_W80_all = number of transmission in which W >= 80%;
 10. n_W80_correctDonorRecipt = number of transmission in point 9 that is a
     true transmission (when comparing to the transmission matrix);
 11. n_W80_swapDonorRecipt = number of transmission in point 9 that is a
     true transmission (when comparing to the transmission matrix), but
     there was a swap between donor and recipient;
 12. expected_n_trans = expected number of transmission defined as the sum
     of all infector probability (W) values.
}
\description{
This function saves several values to compare W calculated on true trees
   (all region and a subset of it)
   and the true transmission matrix. For the return values see below.
}
