% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{track_nodes}
\alias{track_nodes}
\title{Track total number of nodes by atrribute origin}
\usage{
track_nodes(dat, at, total_nodes)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}

\item{total_nodes}{Matrix for total nodes at step at by attribute origin}
}
\value{
the Master list object of network models (\code{dat})
}
\description{
This function will track the total number of nodes in each time step
}
