% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{track_origin}
\alias{track_origin}
\title{Track origin of node in network}
\usage{
track_origin(dat, at, migrations, IDs)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}

\item{migrations}{Migration number from 21 ("global" to "region") or 12
("region" to "global").}

\item{IDs}{Indexes of node IDs}
}
\value{
the Master list object of network models (\code{dat})
}
\description{
This function will track for each node its origin and when it migrates
to another location.
}
