% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{track_stages}
\alias{track_stages}
\title{Track HIV stage of infection}
\usage{
track_stages(dat, HIVstage, at, IDs)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}

\item{IDs}{Indexes of node IDs}

\item{stage}{stage of HIV infection}
}
\value{

}
\description{
This function will track for each node its HIV stage of infection and when
a node change to another stage of infection.
}
