% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_utils.R
\name{add_root_edge}
\alias{add_root_edge}
\title{Add root edge to phylogenetic tree}
\usage{
add_root_edge(tree, total_sim_steps, root.edge_value = 7665)
}
\arguments{
\item{tree}{object of class Phylo}

\item{total_sim_steps}{scalar for the total number of steps of network simulations}

\item{root.edge_value}{scaler for the value to use as root.edge in the phylogenetic tree}
}
\value{

}
\description{
Add root edge to phylogenetic tree
}
