% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.aging.R
\name{aging_msm}
\alias{aging_msm}
\title{Aging Module}
\usage{
aging_msm(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link[EpiModel]{netsim}}.}

\item{at}{Current time step.}
}
\value{
This function returns \code{dat} after updating the nodal attribute
\code{age}.
}
\description{
Module for aging over time for active nodes in the population.
}
